requirements.txt
httpx==0.28.1
beautifulsoup4==4.13.3
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
asyncpg>=0.30.0
psycopg[binary]>=3.2.0
python-dotenv>=1.0.0
